v2.4.18-pre1 -> v2.4.18-pre2
  - APIC LVTERR fixes
  - Fix ppdev ioctl oops and deadlock
  - parport fixes
  - orinoco wireless driver update
  - Fix oopsable race in binfmt_elf.c
  - Small sx16 driver bugfix
  - sbp2 deadlock fix
  - Fix JFFS2 write error handling
  - Intermezzo update
  - Proper AGP support for Intel 830MP chipsets
  - Alpha fixes
  - 53c700 SCSI driver update
  - Fix coredump mmap_sem deadlock on IA64
  - 3ware driver update
  - Fix elevator insertion point on failed
    request merge
  - Remove bogus rpciod_tcp_dispatcher definition
  - Reiserfs fixes
  - de4x5 endianess fixes
  - ISDN CAPI cleanup
  - Make refill_inactive() correctly account
    progress

+++ include/linux/reiserfs_fs_sb.h
@@ -407,7 +407,7 @@ struct reiserfs_sb_info
 				/* To be obsoleted soon by per buffer seals.. -Hans */
     atomic_t s_generation_counter; // increased by one every time the
     // tree gets re-balanced
-    unsigned int s_properties;    /* File system properties. Currently holds
+    unsigned long s_properties;    /* File system properties. Currently holds
 				     on-disk FS format */
     
     /* session statistics */
